♻️ createDashboard 컴포넌트안에 이사시키고 마무리
